requirements.txt
streamlit
python-dotenv
google-generativeai
faiss-cpu
sentence-transformers
pypdf
PyMuPDF
langchain
tiktoken
spacy